squash commit of private branch edit
As long as it does not affect the Fortran source code, multiple
issues reported by markdownlint[1] were fixed.  There however
still are blocks enclosed bt `<small>` and `</small>` Bader
wanted to represent in a different way -- they are preserved.

The workflows as captured in the Makefile were extended.  In part
because the linting is not complete yet, the creation of a pdf
via restructured text (rst) appears as the one requiring the least
manual interventionl.  The rst2pdf however not only scales code
blocks, but large images instead of carrying the png over to the
next page (similar to a floating box in pdfLaTeX).  The issues
with roff (Groff ms) and TeX (pdfLaTeX) are opposite to the ones
in rst2pdf.

[1] https://github.com/markdownlint/markdownlint in version 0.13.0
as packaged for Debian, https://tracker.debian.org/pkg/ruby-mdl