squash commit from branch edit02
The syntax in the markdown file was checked further.  The
automated conversion of the markdownwiki formatted source file
into the markdown format with pandoc allowed a couple of syntax
errors in the snippets of Fortran which were not spotted earlier
because they did not represent an error in the syntax of markdown.

For the preservation of text/Fortran source code, the processing
via rst2pdf is ahead of pdfLaTeX and groff.  pdfLaTeX again is
hampered by a laytout always about to tuncate long lines of
Fortran source code.  The default font family P in the local setup
of groff does not inclde the glyph to express the `mapsto` arrow
in the Bessel function.  On the other hand, rst2pdf continues to
shrink at least one of the two illustrations while rendering the
pdf.

The next steps ahead thus should create/identify Pandoc templates
more appropriate for the format conversions.