use pandoc style template for pdfLaTeX workflow
The additional pandoc style template influences the work of the
pdfengine pdfLaTeX.  With adjustements in the workflow, it is used
using its default parameters.  It is published under the BSD-3 clause
license on https://github.com/Wandmalfarbe/pandoc-latex-template
which permits public use, and edit to adjust it to local needs.