start to edit/adjust eisvogel template
While it is possible to relay variables to Pandoc, e.g. to consider
libertine as the default font family instead of source code pro
set as default within an additional

```
-V fontfamily libertine, libertinust1math, beramono
```

in the call to pandoc, I didn't see how to successfully scale
beramono (to present the source code).  There are a few LaTeX
related issues left (dimension of the illustrations, format of
figure caption, a link color which would be both less intrusive
while reading the document on the screen/retained if print to
paper, etc) not yet addressed.