edit figure caption format
For the first illustration, about the acyclic graph, Bader didn't
prepare a figure caption.  For the second though, he did.  Thus
it seems fine to edit the tex style template to not show a "Fig."
if there isn't a detail to add there.

Perhaps there is a better place in the template to to provide this
edit, though.