use eisvogel's listing option
The style template eisvogel.tex offers to relay the work to LaTeX
usepackage listings.  This then provides small frames around the
snippets of code, and to arrange them without a page break (where
possible); both seen as an improvement to the former representation.
Equally, the intermediate tex source file is easier to read (if
necessary), than the earlier version.

However, the template defaults to Java as the language displayed
in the boxes.  The intended syntax highlighting for Fortran
requires an adjustment.