revert to a Pandoc listing style (now tango)
Pandoc already provides a listing style (tango) at leas very much
similar to the one used in the learning section of fortran-lang.[1]
Hence, in order to keep the pattern here in sync with this reference
and to keep the maintenance at bay, I revert back to one of the
styles already provided by Pandoc (though it isn't its default
previously used).

There might be snippets of Fortran source code starting on one, and
continuing on the next page because these blocks now are not managed
by LaTeX's usepackage listings[2] and the now unavailable process
variable `listings-no-page-break`.  But I think i) the different
font family used in the snippets of Fortran code, ii) indentation,
and iii) their light gray background gently indicate the reader
about the continuation of such a block.  Oxford University Press
offers a preview of the current 6th edition of "Modern Fortran
Explained", and the authors there opted-in for i) and ii), but not
on iii).  Similar for "Modern Fortran in Practice" by Arien Markus
published in 2012.

[1] https://fortran-lang.org/learn/quickstart/
[2] https://ctan.org/pkg/listings