modify Baders <small> </small> block format
A couple of times, Bader's draft included sections enclosed in HTML
<small> and </small> to provide additional insight.  By default,
Pandoc is not able to reliably translate this into other markup
languages (e.g., intermediate tex).  As a temporary fix and because
each of these sections was only a single paragraph, I presume a
brief "Hint:" is a tag not too much of a nuisance to indicate
these as a supplementary piece of information.