test Fortran code blocks which do not pass pagebreaks
This tests a suggestion by M Metcalf to test the generation of the
pdf with pdfLaTeX moderated by pandoc and its style template: how
would the result look like with the additional constraint to not
allow code blocks of Fortran to pass across a page break?

Now, the code blocks reside in a minipage.  If there is too little
remaining space on the current page, this float is relayed to the
next page.  This approach can create larger white space on some
pages (cf. p. 14-15, 21-22, 32-33).