adjust linewrap management
The previous flag `--wrap=none` for Pandoc was good to allow a
longer figure caption, but not for the table in section 8/Performing
I/O with objects of container-like type.  The (implicit) default
linewrap, or its complete drop now is substitituted by an explicit
threshold to break lines longer than 100 characters.

In similar direction is the simultaneous change of the relative
column widths (previous ratio 4:4, now 4:10) in favor of the content
of the second one.

At stage of rst2pdf, there is still a warning about the inheritance
diagram.  While the algorithm claims it were too tall and hence would
rescale it, the eventually compiled pdf squezes it to a size of about
five lines high.  At this scale, it is no longer intelligble.