manually create a mapsto glyph
Dean Allen Provins' "The Groff and Friends HOWTO"[1] (revision
July 2001) describes with the example of the TeX logo how to
construct a glyph missing which here can be applied to substitute
missing \[u21A6] in the current font family to yield a lookalike
of LaTeX's \mapsto.  It took some testing to identify the correct
numbers of escaping backslashes with sed, and the construct might
be not as robust than a glyph properly encoded by design.

[1] https://www.troff.org/TheGroffFriendsHowto.pdf