substitute inheritance diagram (now: inkscape)
The diagram is redrawn from scracth with inkscape for a nicer
representation than the one previously created with draw.io.  The
creation and subsequent management of layers, or the programmatic
rotation of elements (e.g., turn this triangle by 42 degrees) in
inkscape isn't yet as powerful, or available at all in daw.io.