reorganize sections of Bader's draft
This repository's data serve as foundation to add Bader's draft
to the learning section of fortran-lang.org.[1]  It became evident
my rescue of the data from Wikipedia were erroneous when running a

```shell
grep ^= source.mediawiki
```

on the equally retained raw data of branch `upstream`.  This commit
synchronizes the local organization (back) to the way intended by
Reinhold Bader and commit d29f638 of the GitHub repository of the
fortran-lang/webpage.

[1] https://fortran-lang.org/learn/

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>