synchronize with fortran-lang booklet, object_based
In preparation of the material for fortran-lang.org, hints for a
more contemporary layout of the snippets of code was recommended.
This commit basically copy-pasts the results from there into the
local repository for the benefit of the later: consistent 2 spaces
per level of indentation, lower-case spelling of keywords anyway
subject to syntax highlighting.

Contrasting to the submission to fortran-lang.org, the snippets of
code retain their label `f90` provided they display functional Fortran
source code, or the the mere principle only (i.e., ellipsis `...` and
`:` colon are place holders for specific code to be added).  Second,
hints still are represented in the style supported by Pandoc.

Because the source code of the booklet for fortran-lang.org was split
into smaller markdown files, this incremental  update of `pandoc_md.md`
considers the material up to and including the first chapter (file
`object_based_programming_techniques.md`).

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>